+++ src/main/java/com/example/demo/src/user/model/PatchUserReq.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class PatchUserReq {
-    private int userIdx;
+    private Long userIdx;
     private String userName;
 }
